Disable checking of ‘required options’
Since the program configuration can now be read from the main application’s preferences, the complexity of determining which options are required and unspecified has increased.

This validation is being removed since the only options that are marked as required are the music media dir + output path args which already have individual extensive checks.